requirements.txt
# Core dependencies for Multi-Agent Research System
# Phase 1-2: Web Search Tool + Researcher Agent

# Async HTTP client (for API calls and web search)
aiohttp>=3.9.0

# DuckDuckGo search (free option, no API key needed)
ddgs>=4.0.0

# Environment variable management
python-dotenv>=1.0.0

# Data validation and structured outputs
pydantic>=2.5.0

# Anthropic Claude API (for agents)
anthropic>=0.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality (optional)
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0